(PC-40074)[PRO] refactor: rename new homepage files and components
